Move entity code into feature package
