--- src/main/kotlin/org/coner/snoozle/db/Entity.kt
@@ -1,3 +1,0 @@
-package org.coner.snoozle.db
-
-interface Entity
