+++ src/main/kotlin/org/coner/snoozle/db/entity/CurrentVersionRecord.kt
@@ -1,4 +1,4 @@
-package org.coner.snoozle.db
+package org.coner.snoozle.db.entity
 
 import java.time.ZonedDateTime
