+++ src/main/kotlin/org/coner/snoozle/db/entity/Entity.kt
@@ -1,0 +1,3 @@
+package org.coner.snoozle.db.entity
+
+interface Entity
