+++ src/main/kotlin/org/coner/snoozle/db/entity/EntityResource.kt
@@ -1,10 +1,9 @@
-package org.coner.snoozle.db
+package org.coner.snoozle.db.entity
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import de.helmbold.rxfilewatcher.PathObservables
 import io.reactivex.Observable
 import org.coner.snoozle.db.path.Pathfinder
-import org.coner.snoozle.util.hasUuidPattern
 import org.coner.snoozle.util.nameWithoutExtension
 import org.coner.snoozle.util.uuid
 import java.nio.file.Files
