+++ src/main/kotlin/org/coner/snoozle/db/entity/IoDelegate.kt
@@ -1,4 +1,4 @@
-package org.coner.snoozle.db
+package org.coner.snoozle.db.entity
 
 internal interface IoDelegate<E : Entity> {
