+++ src/test/kotlin/org/coner/snoozle/db/sample/Widget.kt
@@ -1,6 +1,6 @@
 package org.coner.snoozle.db.sample
 
-import org.coner.snoozle.db.Entity
+import org.coner.snoozle.db.entity.Entity
 import java.util.*
 
 data class Widget(
